requirements.txt

requests
numpy

streamlit
pandas
google-cloud-bigquery
google-auth
db-dtypes
plotly